fix hitbtc place order

+++ xchange-hitbtc/src/main/java/org/knowm/xchange/hitbtc/HitbtcAuthenticated.java
@@ -45,7 +45,7 @@ public interface HitbtcAuthenticated extends Hitbtc {
   public HitbtcExecutionReportResponse postHitbtcNewOrder(@HeaderParam("X-Signature") ParamsDigest signature,
       @QueryParam("nonce") SynchronizedValueFactory<Long> valueFactory, @QueryParam("apikey") String apiKey,
       @FormParam("clientOrderId") String clientOrderId, @FormParam("symbol") String symbol, @FormParam("side") String side,
-      @FormParam("price") BigDecimal price, @FormParam("quantity") BigInteger quantity, // 1 lot = 0.01 BTC
+      @FormParam("price") BigDecimal price, @FormParam("quantity") BigDecimal quantity, // 1 lot = 0.01 BTC
       @FormParam("type") String type, @FormParam("timeInForce") String timeInForce) throws IOException, HitbtcException;
 
   @POST
